mcm c++ tests moved from Qubit to QubitManaged
